Refactor REST API to use WP_REST_Controller pattern
- Add rest/ directory with Endpoint_Controller and Media_Controller
- Controllers extend WP_REST_Controller and use Micropub trait
- Add separate Discovery class for endpoint discovery
- Update Micropub singleton to properly hook into rest_api_init, init, admin_notices
- Add MICROPUB_PLUGIN_VERSION constant
- Remove old Base, Endpoint, and Media classes
