+++ includes/functions.php
@@ -308,3 +308,48 @@ if ( ! function_exists( 'micropub_wp_error' ) ) {
 		return null;
 	}
 }
+
+if ( ! function_exists( 'micropub_get_mf2' ) ) {
+	/**
+	 * Get MF2 properties for a post.
+	 *
+	 * @param int|null $post_id Post ID.
+	 * @return array MF2 properties.
+	 */
+	function micropub_get_mf2( $post_id = null ) {
+		$mf2  = array();
+		$post = get_post( $post_id );
+
+		foreach ( get_post_meta( $post_id ) as $field => $val ) {
+			$val = maybe_unserialize( $val[0] );
+			if ( 'mf2_type' === $field ) {
+				$mf2['type'] = $val;
+			} elseif ( 'mf2_' === substr( $field, 0, 4 ) ) {
+				$mf2['properties'][ substr( $field, 4 ) ] = $val;
+			}
+		}
+
+		// Time Information.
+		$published                      = micropub_get_post_datetime( $post );
+		$updated                        = micropub_get_post_datetime( $post, 'modified' );
+		$mf2['properties']['published'] = array( $published->format( DATE_W3C ) );
+
+		if ( $published->getTimestamp() !== $updated->getTimestamp() ) {
+			$mf2['properties']['updated'] = array( $updated->format( DATE_W3C ) );
+		}
+
+		if ( ! empty( $post->post_title ) ) {
+			$mf2['properties']['name'] = array( $post->post_title );
+		}
+
+		if ( ! empty( $post->post_excerpt ) ) {
+			$mf2['properties']['summary'] = array( htmlspecialchars_decode( $post->post_excerpt ) );
+		}
+
+		if ( ! array_key_exists( 'content', $mf2['properties'] ) && ! empty( $post->post_content ) ) {
+			$mf2['properties']['content'] = array( htmlspecialchars_decode( $post->post_content ) );
+		}
+
+		return $mf2;
+	}
+}
